+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Namespace component classes with `sb`.
+
 ## [v0.1.1] - 2022-10-02
 
 - Testing deployment.