+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 
 - Namespace component and main classes with `sb`.
+- Add Icon, Tabs, and Textarea components.
+- Add Colors utility.
+- Improve documentation.
 
 ## [v0.1.1] - 2022-10-02