+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.2.0] - 2022-10-02
+
 - Namespace component and main classes with `sb`.
 - Add Icon, Modal, Tabs, and Textarea components.
 - Add Colors utility.